# dnsproxy: introduce cache to reference count regexes