--- a/pkg/fqdn/dnsproxy/proxy.go
+++ b/pkg/fqdn/dnsproxy/proxy.go
@@ -138,7 +138,6 @@ type DNSProxy struct {
 	// allowed tracks all allowed L7 DNS rules by endpointID, destination port,
 	// and L3 Selector. All must match for a query to be allowed.
 	//
-	// matchNames with no regexp wildcards are still compiled, internally.
 	// Note: Simple DNS names, e.g. bar.foo.com, will treat the "." as a literal.
 	allowed perEPAllow
 
@@ -146,6 +145,11 @@ type DNSProxy struct {
 	// used until 'allowed' rules for an endpoint are first initialized after
 	// a restart
 	restored perEPRestored
+
+	// cache is an internal structure to keep track of all the in use DNS rules. We do that
+	// so that we avoid storing multiple similar versions of the same rules, so that we can improve
+	// performance and reduce memory consumption when multiple endpoints or ports have similar rules.
+	cache regexCache
 
 	// mapping restored endpoint IP (both IPv4 and IPv6) to *Endpoint
 	restoredEPs restoredEPs
@@ -158,6 +162,17 @@ type DNSProxy struct {
 	unbindAddress func()
 }
 
+// regexCacheEntry is a lookup entry used to cache a compiled regex
+// and how many references it has
+type regexCacheEntry struct {
+	regex          *regexp.Regexp
+	referenceCount int
+}
+
+// regexCache is a reference counted cache used for reusing the compiled regex when multiple policies
+// have the same set of rules, or the same rule applies to multiple endpoints.
+type regexCache map[string]*regexCacheEntry
+
 // perEPAllow maps EndpointIDs to ports + selectors + rules
 type perEPAllow map[uint64]portToSelectorAllow
 
@@ -165,14 +180,27 @@ type perEPAllow map[uint64]portToSelectorAllow
 type portToSelectorAllow map[uint16]CachedSelectorREEntry
 
 // CachedSelectorREEntry maps port numbers to selectors to rules, mirroring
-// policy.L7DataMap but the DNS rules are compiled into a single regexp
+// policy.L7DataMap but the DNS rules are compiled into a regex
 type CachedSelectorREEntry map[policy.CachedSelector]*regexp.Regexp
 
 // structure for restored rules that can be used while Cilium agent is restoring endpoints
 type perEPRestored map[uint64]restore.DNSRules
 
+// restoredIPRule is the dnsproxy internal way of representing a restored IPRule
+// where we also store the actual compiled regular expression as a, as well
+// as the original restored IPRule
+type restoredIPRule struct {
+	restore.IPRule
+	regex *regexp.Regexp
+}
+
 // map from EP IPs to *Endpoint
 type restoredEPs map[string]*endpoint.Endpoint
+
+// asIPRule returns a new restore.IPRule representing the rules, including the provided IP map.
+func asIPRule(r *regexp.Regexp, IPs map[string]struct{}) restore.IPRule {
+	return restore.IPRule{IPs: IPs, Re: restore.RuleRegex{Regexp: r}}
+}
 
 // CheckRestored checks endpointID, destPort, destIP, and name against the restored rules,
 // and only returns true if a restored rule matches.
@@ -230,7 +258,7 @@ func (p *DNSProxy) GetRules(endpointID uint16) (restore.DNSRules, error) {
 					}
 				}
 			}
-			ipRules = append(ipRules, restore.IPRule{IPs: IPs, Re: restore.RuleRegex{Regexp: regex}})
+			ipRules = append(ipRules, asIPRule(regex, IPs))
 		}
 		restored[port] = ipRules
 	}
@@ -275,55 +303,133 @@ func (p *DNSProxy) RemoveRestoredRules(endpointID uint16) {
 	p.removeRestoredRulesLocked(uint64(endpointID))
 }
 
-// setPortRulesForID sets the matching rules for endpointID and destPort for
-// later lookups. It converts newRules into a unified regexp that can be reused
-// later.
-func (allow perEPAllow) setPortRulesForID(endpointID uint64, destPort uint16, newRules policy.L7DataMap) error {
-	// This is the delete case
-	if len(newRules) == 0 {
-		epPorts := allow[endpointID]
-		delete(epPorts, destPort)
-		if len(epPorts) == 0 {
-			delete(allow, endpointID)
-		}
+// lookupOrCompileRegex will check if the pattern is already compiled and present in another policy, and
+// will reuse it in order to reduce memory consumption. The usage is reference counted, so all calls where
+// lookupOrCompileRegex returns no error, a subsequent call to release it via releaseRegex has to
+// be done when it's no longer being used by the policy.
+func (c regexCache) lookupOrCompileRegex(pattern string) (*regexp.Regexp, error) {
+	if entry, ok := c[pattern]; ok {
+		entry.referenceCount += 1
+		return entry.regex, nil
+	}
+	regex, err := re.CompileRegex(pattern)
+	if err != nil {
+		return nil, err
+	}
+	c[pattern] = &regexCacheEntry{regex: regex, referenceCount: 1}
+	return regex, nil
+}
+
+// lookupOrInsertRegex is equivalent to lookupOrCompileRegex, but a compiled regex is provided
+// instead of the pattern. In case a compiled regex with the same pattern as the provided regex is already present in
+// the cache, the already present regex will be returned. By doing that, the duplicate can be garbage collected in case
+// there are no other references to it. Trying to insert a nil value is a noop and will return nil
+func (c regexCache) lookupOrInsertRegex(regex *regexp.Regexp) *regexp.Regexp {
+	if regex == nil {
 		return nil
 	}
+	pattern := regex.String()
+	if entry, ok := c[pattern]; ok {
+		entry.referenceCount += 1
+		return entry.regex
+	}
+	c[pattern] = &regexCacheEntry{regex: regex, referenceCount: 1}
+	return regex
+}
 
-	newRE, err := GetSelectorRegexMap(newRules)
+// releaseRegex releases the provided regex. In case there are no longer any references to it,
+// it will be freed. Running release on a nil value is a noop.
+func (c regexCache) releaseRegex(regex *regexp.Regexp) {
+	if regex == nil {
+		return
+	}
+	pattern := regex.String()
+	if indexEntry, ok := c[pattern]; ok {
+		switch indexEntry.referenceCount {
+		case 1:
+			delete(c, pattern)
+		default:
+			indexEntry.referenceCount -= 1
+		}
+	}
+}
+
+// removeAndReleasePortRulesForID removes the old port rules for the given destPort on the given endpointID. It also
+// releases the regexes so that unused regex can be freed from memory.
+func (allow perEPAllow) removeAndReleasePortRulesForID(cache regexCache, endpointID uint64, destPort uint16) {
+	epPorts, hasEpPorts := allow[endpointID]
+	if !hasEpPorts {
+		return
+	}
+	for _, m := range epPorts[destPort] {
+		cache.releaseRegex(m)
+	}
+	delete(epPorts, destPort)
+	if len(epPorts) == 0 {
+		delete(allow, endpointID)
+	}
+}
+
+// setPortRulesForID sets the matching rules for endpointID and destPort for
+// later lookups. It converts newRules into a compiled regex
+func (allow perEPAllow) setPortRulesForID(cache regexCache, endpointID uint64, destPort uint16, newRules policy.L7DataMap) error {
+	if len(newRules) == 0 {
+		allow.removeAndReleasePortRulesForID(cache, endpointID, destPort)
+		return nil
+	}
+	cse := make(CachedSelectorREEntry, len(newRules))
+	var err error
+	for selector, newRuleset := range newRules {
+		pattern := GeneratePattern(newRuleset)
+
+		var regex *regexp.Regexp
+		regex, err = cache.lookupOrCompileRegex(pattern)
+		if err != nil {
+			break
+		}
+		cse[selector] = regex
+	}
 	if err != nil {
+		// Unregister the registered regexes before returning the error to avoid
+		// leaving unused references in the cache
+		for k, regex := range cse {
+			cache.releaseRegex(regex)
+			delete(cse, k)
+		}
 		return err
 	}
-
+	allow.removeAndReleasePortRulesForID(cache, endpointID, destPort)
 	epPorts, exist := allow[endpointID]
 	if !exist {
 		epPorts = make(portToSelectorAllow)
 		allow[endpointID] = epPorts
 	}
-
-	epPorts[destPort] = newRE
+	epPorts[destPort] = cse
 	return nil
 }
 
 // setPortRulesForIDFromUnifiedFormat sets the matching rules for endpointID and destPort for
-// later lookups.
-func (allow perEPAllow) setPortRulesForIDFromUnifiedFormat(endpointID uint64, destPort uint16, newRules CachedSelectorREEntry) error {
-	// This is the delete case
+// later lookups. It does not guarantee it will reuse all the provided regexes, since it will reuse
+// already existing regexes with the same pattern in case they are already in use.
+func (allow perEPAllow) setPortRulesForIDFromUnifiedFormat(cache regexCache, endpointID uint64, destPort uint16, newRules CachedSelectorREEntry) error {
 	if len(newRules) == 0 {
-		epPorts := allow[endpointID]
-		delete(epPorts, destPort)
-		if len(epPorts) == 0 {
-			delete(allow, endpointID)
-		}
+		allow.removeAndReleasePortRulesForID(cache, endpointID, destPort)
 		return nil
 	}
+	cse := make(CachedSelectorREEntry, len(newRules))
+	for selector, providedRegex := range newRules {
+		// In case the regex is already compiled and in use in another regex, lookupOrInsertRegex
+		// will return a ref. to the existing regex, and use that one.
+		cse[selector] = cache.lookupOrInsertRegex(providedRegex)
+	}
 
+	allow.removeAndReleasePortRulesForID(cache, endpointID, destPort)
 	epPorts, exist := allow[endpointID]
 	if !exist {
 		epPorts = make(portToSelectorAllow)
 		allow[endpointID] = epPorts
 	}
-
-	epPorts[destPort] = newRules
+	epPorts[destPort] = cse
 	return nil
 }
 
@@ -458,6 +564,7 @@ func StartDNSProxy(
 		allowed:                  make(perEPAllow),
 		restored:                 make(perEPRestored),
 		restoredEPs:              make(restoredEPs),
+		cache:                    make(regexCache),
 		EnableDNSCompression:     enableDNSCompression,
 		maxIPsPerRestoredDNSRule: maxRestoreDNSIPs,
 	}
@@ -546,7 +653,7 @@ func (p *DNSProxy) UpdateAllowed(endpointID uint64, destPort uint16, newRules po
 	p.Lock()
 	defer p.Unlock()
 
-	err := p.allowed.setPortRulesForID(endpointID, destPort, newRules)
+	err := p.allowed.setPortRulesForID(p.cache, endpointID, destPort, newRules)
 	if err == nil {
 		// Rules were updated based on policy, remove restored rules
 		p.removeRestoredRulesLocked(endpointID)
@@ -559,7 +666,7 @@ func (p *DNSProxy) UpdateAllowedFromSelectorRegexes(endpointID uint64, destPort 
 	p.Lock()
 	defer p.Unlock()
 
-	err := p.allowed.setPortRulesForIDFromUnifiedFormat(endpointID, destPort, newRules)
+	err := p.allowed.setPortRulesForIDFromUnifiedFormat(p.cache, endpointID, destPort, newRules)
 	if err == nil {
 		// Rules were updated based on policy, remove restored rules
 		p.removeRestoredRulesLocked(endpointID)
@@ -580,9 +687,9 @@ func (p *DNSProxy) CheckAllowed(endpointID uint64, destPort uint16, destID ident
 		return p.checkRestored(endpointID, destPort, destIP.String(), name), nil
 	}
 
-	for selector, re := range epAllow {
+	for selector, regex := range epAllow {
 		// The port was matched in getPortRulesForID, above.
-		if selector.Selects(destID) && re.MatchString(name) {
+		if regex != nil && selector.Selects(destID) && regex.MatchString(name) {
 			return true, nil
 		}
 	}
@@ -978,34 +1085,26 @@ func shouldCompressResponse(request, response *dns.Msg) bool {
 	return false
 }
 
-func GetSelectorRegexMap(l7 policy.L7DataMap) (CachedSelectorREEntry, error) {
-	newRE := make(CachedSelectorREEntry)
-	for selector, l7Rules := range l7 {
-		if l7Rules == nil {
-			l7Rules = &policy.PerSelectorPolicy{L7Rules: api.L7Rules{DNS: []api.PortRuleDNS{{MatchPattern: "*"}}}}
+// GeneratePattern takes a set of l7Rules and returns a regular expression pattern for matching the
+// provided l7 rules.
+func GeneratePattern(l7Rules *policy.PerSelectorPolicy) (pattern string) {
+	if l7Rules == nil {
+		l7Rules = &policy.PerSelectorPolicy{L7Rules: api.L7Rules{DNS: []api.PortRuleDNS{{MatchPattern: "*"}}}}
+	}
+	reStrings := make([]string, 0, len(l7Rules.DNS))
+	for _, dnsRule := range l7Rules.DNS {
+		if len(dnsRule.MatchName) > 0 {
+			dnsRuleName := strings.ToLower(dns.Fqdn(dnsRule.MatchName))
+			reStrings = append(reStrings, "("+matchpattern.ToRegexp(dnsRuleName)+")")
 		}
-		reStrings := make([]string, 0, len(l7Rules.DNS))
-		for _, dnsRule := range l7Rules.DNS {
-			if len(dnsRule.MatchName) > 0 {
-				dnsRuleName := strings.ToLower(dns.Fqdn(dnsRule.MatchName))
-				dnsPatternAsRE := matchpattern.ToRegexp(dnsRuleName)
-				reStrings = append(reStrings, "("+dnsPatternAsRE+")")
-			}
-			if len(dnsRule.MatchPattern) > 0 {
-				dnsPattern := matchpattern.Sanitize(dnsRule.MatchPattern)
-				dnsPatternAsRE := matchpattern.ToRegexp(dnsPattern)
-				reStrings = append(reStrings, "("+dnsPatternAsRE+")")
-			}
+		if len(dnsRule.MatchPattern) > 0 {
+			dnsPattern := matchpattern.Sanitize(dnsRule.MatchPattern)
+			dnsPatternAsRE := matchpattern.ToRegexp(dnsPattern)
+			reStrings = append(reStrings, "("+dnsPatternAsRE+")")
 		}
-		mp := strings.Join(reStrings, "|")
-		rei, err := re.CompileRegex(mp)
-		if err != nil {
-			return nil, err
-		}
-		newRE[selector] = rei
 	}
 
-	return newRE, nil
+	return strings.Join(reStrings, "|")
 }
 
 func (p *DNSProxy) Cleanup() {