@odinuge 2022-09-13
A step towards https://github.com/cilium/cilium/pull/20396 and https://github.com/cilium/cilium/pull/20246, that is smaller and easier to reason around.