@odinuge 2022-10-06
👋 Pushed a few changes now based on your comments Chris. Pushed them as `--fixup` commits so that its easier to track the diff, and then I'll squash them when things are ready!

I'll give it another look tomorrow morning, and maybe add a few more tests of the edge-cases.

Thanks
Odin