@odinuge 2022-10-10
I did find another edge case btw. where `len(l7Rules.DNS) == 0` i changed the behavior to block all traffic instead of allowing all traffic as the baseline/master does. Pushed this diff to keep the current behavior from master; https://github.com/cilium/cilium/compare/492e9cc5c08f0e9fca8ccfe6ae762c13bbbd1284..d6a38a28c9176b5d0e769b33ef5e250107f86713

That case ^ results in an empty regex (the pattern is an empty string), and an empty regex match all input (the parsed syntax is actually `(?:)` in go). That is possibly another optimization we can do to check if the pattern is empty, and in that case, define it as a "always allow" rule we don't need to run `MatchString` on; but thats for another PR.