@christarazi 2022-10-12
From what I can tell, yes these are flakes. Tracking issues:

* https://github.com/cilium/cilium/issues/21710
* https://github.com/cilium/cilium/issues/16203