@odinuge 2022-10-26
Hi @christarazi @ti-mo! Are you waiting for me here, or? Happy to engage if there is anything you want me to do! Would be great to get an agreement here so we could start pushing the actual regex improvements to reduce the heap size and allocations/cpu usage during compilation. 😄