@christarazi 2022-10-26
@odinuge Given that this is a user-facing change, could you write the release note from the user's point of view and the sort of impact that they will see? I'm thinking something like "reduced memory usage with FQDN policies" but with more elaboration.