@christarazi 2022-11-14
@odinuge I don't think there's much left to do here. I think the code looks good. You have tested & validated it your environments which is great. Pending another CI pass, I think it should be ready to merge.

One point for follow up is where does this leave us with the FQDN LRU? Could we simply deprecate it and remove it in the next cycle (1.14)?