@odinuge 2022-11-17
> @odinuge I don't think there's much left to do here. I think the code looks good. You have tested & validated it your environments which is great. Pending another CI pass, I think it should be ready to merge.

Great!

> One point for follow up is where does this leave us with the FQDN LRU? Could we simply deprecate it and remove it in the next cycle (1.14)?

Yeah, that is a good question. I vote for at least removing the lru from the dnsproxy since we now (with this) have a separate cache; but let's do it in a follow up after this.

The other places it might be useful, since we reduce the actual time spent during regex compilation; but haven't digged too deep into those yet.

One thing I do vote for is to remove [this](https://github.com/cilium/cilium/blob/master/pkg/policy/api/fqdn.go#L91) and [this](https://github.com/cilium/cilium/blob/master/pkg/policy/api/fqdn.go#L123) since we already validate that the `matchPattern` is valid. Unless there is some combination of characters that can lead to regex compilation failing when it matches the `allowedPatternChars` regex above?