@christarazi 2022-11-17
* 5.4 hit https://github.com/cilium/cilium/issues/21735. 
* External workloads failing due to provisioning GKE cluster.

Given these are flakes, I think this is ready to merge.